refactor(core): drop split compat barrels, import concrete modules
Delete the five re-export barrels left by the domain regroup
(analysis/intraday.ts, analysis/candlePatterns.ts, ai/lobehub/gateway.ts,
ai/agents/agentTools.ts, ai/personas/analyst.ts) and repoint every consumer
at the concrete submodule that declares each symbol.
