--- packages/core/src/ai/lobehub/gateway.ts
@@ -1,1 +1,0 @@
-export { WebApiLobeHubCloudGateway } from './gateway/client.js';
