--- packages/core/src/analysis/candlePatterns.ts
@@ -1,2 +1,0 @@
-export { detectCandlePatterns } from './candlePatterns/detect.js';
-export { CANDLE_PATTERN_META } from './candlePatterns/meta.js';
