+++ packages/core/src/pro/aiExtension.ts
@@ -1,6 +1,6 @@
 import type { AgentMessage } from '@earendil-works/pi-agent-core';
 import type { ProAiTranscriptMessage, ProAiTurnContext } from '@kansoku/pro-api';
-import type { FsReadMount } from '../ai/agents/agentTools.js';
+import type { FsReadMount } from '../ai/agents/agentTools/fsMounts.js';
 import { textOf } from '../ai/conversation/conversationShared.js';
 import { BaseFirstUserContentProvider } from '../ai/conversation/messages/injectors/baseFirstUserContentProvider.js';
 import {
