+++ packages/core/src/realtime/aiChannels.ts
@@ -1,6 +1,6 @@
 import type { ProChannel } from '@kansoku/pro-api';
 import type { CockpitComment } from '@kansoku/shared/types';
-import { listAnalystRuns, onAnalystRunChange } from '../ai/personas/analyst.js';
+import { listAnalystRuns, onAnalystRunChange } from '../ai/personas/analyst/runState.js';
 import { assistantChatTurnState, onAssistantChatEvent } from '../ai/assistant/assistantChat.js';
 import { type ChatEvent, chatTurnState, onChatEvent } from '../ai/chat/chat.js';
 import { listComments, onAnyComment, onComment } from '../ai/personas/comments.js';
