+++ packages/core/test/candlePatterns.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { detectCandlePatterns } from '../src/analysis/candlePatterns.js';
+import { detectCandlePatterns } from '../src/analysis/candlePatterns/detect.js';
 import { loadFixture } from './helpers.js';
 
 interface TestBar {
