+++ packages/core/test/channelProtocol.test.ts
@@ -5,7 +5,8 @@ import type { AgentEvent } from '@earendil-works/pi-agent-core';
 import type { AiAgentFactory, AiAgentHandle } from '../src/ai/agents/agentSession.js';
 import type { AiModel } from '../src/ai/runtime/models.js';
 import { runAssistantChatTurn } from '../src/ai/assistant/assistantChat.js';
-import { type AnalystDeps, runAnalyst } from '../src/ai/personas/analyst.js';
+import type { AnalystDeps } from '../src/ai/personas/analyst/types.js';
+import { runAnalyst } from '../src/ai/personas/analyst/run.js';
 import { createAssistantSession } from '../src/ai/assistant/assistantChatStore.js';
 import type { ReassessPack } from '../src/ai/agents/datapack.js';
 import { createDb, type Db } from '../src/db/index.js';
