+++ packages/core/test/intradayOffSessionMarket.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 import type { RawBar } from '@kansoku/shared/types';
-import { buildIntraday, type IntradayInput } from '../src/analysis/intraday.js';
+import { buildIntraday, type IntradayInput } from '../src/analysis/intraday/orchestrator.js';
 
 const DAY = { y: 2026, m: 6, d: 8 };
